Create ajax-hustle-bunny.js
This file is not ready for use. I just pulled it out of the project that it was created to help handle Ajax. So I yanked a few project specific lines from it and in a few weeks when this project is over I will come back and clean out the stand alone file to get it ready for use by anyone. At that time I will fix any bugs and add a README.md