Update ajax-hustle-bunny.js
Updated much of the file. I need to go through this and add in "use strict"; and make some code more solid in certain areas. When I began this creation I didn't think that it would become as useful as it is. So I want to make it more robust.